requirements.txt
apache-airflow==2.4.2

astro-sdk-python==1.2.0

apache-airflow-providers-microsoft-azure==4.3.0
apache-airflow-providers-amazon==6.0.0
apache-airflow-providers-google==8.4.0

pytest==7.2.0
openlineage-python==0.15.1
openlineage-airflow==0.15.1